consensus path db automated update, get_pathways()

+++ codes3d/pathways.py
@@ -4,43 +4,42 @@
 
 from __future__ import print_function
 from sys import argv
-from sys import exit
 import os
 from StringIO import StringIO
 from pprint import PrettyPrinter
 from json import loads
-from csv import reader
 import sqlite3
 from pycurl import Curl
 from pycurl import HTTP_CODE
 import configparser
+from requests import post
 from wikipathways_api_client import WikipathwaysApiClient
 
 def wikipathways(gene, exclude_reactome=True):
-    """WikiPathways"""
+    """Wikipathways"""
     wp_client = WikipathwaysApiClient()
     kwargs = {'query': gene,
               'organism': 'http://identifiers.org/taxonomy/9606'
              }
 
-    pathways = {pathway['identifier']: pathway['name']
+    pathways = {('Wikipathways', pathway['identifier']): pathway['name']
                 for pathway in wp_client.find_pathways_by_text(**kwargs)}
 
     if not exclude_reactome:
         return pathways
 
-    wp_base_url = 'https://www.wikipathways.org/index.php'
+    base_url = 'https://www.wikipathways.org/index.php'
     client = Curl()
     query_buffer = StringIO()
     unique_pathways = {}
     for pathway_id in pathways:
-        client.setopt(client.URL, '%s/Pathway:%s' % (wp_base_url, pathway_id))
+        client.setopt(client.URL, '%s/Pathway:%s' % (base_url, pathway_id[1]))
         client.setopt(client.WRITEDATA, query_buffer)
         client.perform()
 
-        wp_response = query_buffer.getvalue()
+        response = query_buffer.getvalue()
         query_buffer.truncate(0)
-        if 'http://www.reactome.org/PathwayBrowser/#DIAGRAM=' not in wp_response:
+        if 'http://www.reactome.org/PathwayBrowser/#DIAGRAM=' not in response:
             unique_pathways[pathway_id] = pathways[pathway_id]
 
     return unique_pathways
@@ -84,41 +83,36 @@ def reactome(gene):
             pathway_response = loads(query_buffer.getvalue())
             query_buffer.truncate(0)
             for pathway in pathway_response:
-                pathways[pathway['stId'].encode('utf-8')] = pathway[
+                pathways[('Reactome', pathway['stId'].encode('utf-8'))] = pathway[
                     'displayName'].encode('utf-8')
 
     return pathways
 
 def reactome_in_wikipathways(gene):
     """Identify pathways present in WikiPathways and Reactome"""
-    wp_base_url = 'https://www.wikipathways.org/index.php'
+    base_url = 'https://www.wikipathways.org/index.php'
     query_buffer = StringIO()
     client = Curl()
 
     pathways = wikipathways(gene, False)
     common_pathways = {}
     for pathway_id in pathways:
-        client.setopt(client.URL, '%s/Pathway:%s' % (wp_base_url, pathway_id))
+        client.setopt(client.URL, '%s/Pathway:%s' % (base_url, pathway_id[1]))
         client.setopt(client.WRITEDATA, query_buffer)
         client.perform()
-        wp_response = query_buffer.getvalue()
+        response = query_buffer.getvalue()
         query_buffer.truncate(0)
 
-        if 'http://www.reactome.org/PathwayBrowser/#DIAGRAM=' not in wp_response:
+        if 'http://www.reactome.org/PathwayBrowser/#DIAGRAM=' not in response:
             continue
 
-        position = wp_response.find('http://www.reactome.org/PathwayBrowser/#DIAGRAM=')
-        position += 48
+        position = 48 + response.find('http://www.reactome.org/PathwayBrowser/#DIAGRAM=')
         reactome_id = 'R-HSA-'
-        while wp_response[position].isdigit():
-            reactome_id += wp_response[position]
+        while response[position].isdigit():
+            reactome_id += response[position]
             position += 1
 
-        try:
-            common_pathways[reactome_id][pathway_id] = pathways[pathway_id]
-        except KeyError:
-            common_pathways[reactome_id] = {}
-            common_pathways[reactome_id][pathway_id] = pathways[pathway_id]
+        common_pathways[('Reactome', reactome_id)] = pathways[pathway_id]
 
     return common_pathways
 
@@ -134,6 +128,7 @@ def kegg(gene):
     client.perform()
     gene_entries = [entry for entry in
                     query_buffer.getvalue().split('\n') if entry]
+    query_buffer.truncate(0)
 
     if not gene_entries:
         return {}
@@ -141,8 +136,7 @@ def kegg(gene):
     gene_ids = []
     for entry in gene_entries:
         entry = entry.split('\t')
-        if  entry[0].startswith('hsa:') and gene in entry[1].split(
-                ';')[0].split(','):
+        if  entry[0].startswith('hsa:') and gene in entry[1].split(';')[0].split(','):
             gene_ids.append(entry[0])
 
     if not gene_ids:
@@ -150,50 +144,48 @@ def kegg(gene):
 
     pathways = {}
     for gene_id in gene_ids:
-        query_buffer.truncate(0)
         client.setopt(client.URL, '%s/link/pathway/%s' % (api_url, gene_id))
         client.setopt(client.WRITEDATA, query_buffer)
         client.perform()
         pathway_ids = [entry.split('\t')[1] for entry
                        in query_buffer.getvalue().split('\n') if entry]
+        query_buffer.truncate(0)
 
         if not pathway_ids:
             continue
 
         for pathway_id in pathway_ids:
-            query_buffer.truncate(0)
             client.setopt(client.URL, '%s/get/%s' % (api_url, pathway_id))
             client.setopt(client.WRITEDATA, query_buffer)
             client.perform()
             pathway_response = query_buffer.getvalue().split('\n')
             pathway_name = pathway_response[1][5:-23].strip()
-            pathways[pathway_id] = pathway_name
+            pathways[('KEGG', pathway_id)] = pathway_name
 
-    query_buffer.truncate(0)
     return pathways
 
-def build_consensus_db():
-    """Build, respectively update local ConsensusDB database"""
+def build_consensuspathdb():
+    """Build inverted local ConsensusPathDB database"""
+    url = 'http://cpdb.molgen.mpg.de/CPDB/getPathwayGenes'
     config = configparser.ConfigParser()
-    config.read(os.path.join(os.path.dirname(__file__),  '../docs/codes3d.conf'))
-    tab_dir = os.path.join(os.path.dirname(__file__), config.get('Defaults', 'CONSENSUS_DB_TAB_FP'))
+    config.read(os.path.join(os.path.dirname(__file__), '../docs/codes3d.conf'))
     db_dir = os.path.join(os.path.dirname(__file__), config.get('Defaults', 'CONSENSUS_DB_FP'))
 
-    try:
-        pathway_file = open(tab_dir, 'r')
-    except IOError:
-        print('%s could not be updated.' % db_dir)
-        exit()
+    file_request = post(url, data={'idtype': 'hgnc-symbol'})
+    if file_request.status_code != 200:
+        return False
 
     db_connect = sqlite3.connect(db_dir)
     db_cursor = db_connect.cursor()
     db_cursor.execute("""DROP TABLE IF EXISTS pathways""")
-    db_cursor.execute("""CREATE TABLE pathways (database text, gene text, pathway_id text, pathway_label text)""")
+    db_cursor.execute("""CREATE TABLE pathways
+                         (database text, gene text, pathway_id text, pathway_label text)""")
 
-    tsv_reader = reader(pathway_file, delimiter='\t')
-    next(tsv_reader, None)
+    pathway_data = file_request.content.split('\n')
+    del pathway_data[0]
 
-    for line in tsv_reader:
+    for line in pathway_data:
+        line = line.split('\t')
         pathway_label = unicode(line[0], 'utf-8')
         pathway_id = unicode(line[1], 'utf-8')
         database = unicode(line[2], 'utf-8')
@@ -207,56 +199,67 @@ def build_consensus_db():
 
     db_connect.commit()
     db_connect.close()
-    print('%s successfully built.' % db_dir)
 
-def consensus_db(gene):
-    """ConsensusDB"""
+    return True
+
+def remove_consensuspathdb():
+    """Remove inverted local ConsensusPathDB database"""
+    config = configparser.ConfigParser()
+    config.read(os.path.join(os.path.dirname(__file__), '../docs/codes3d.conf'))
+    db_dir = os.path.join(os.path.dirname(__file__), config.get('Defaults', 'CONSENSUS_DB_FP'))
+    try:
+        os.remove(db_dir)
+    except OSError:
+        return False
+    return True
+
+def consensuspathdb(gene):
+    """ConsensusPathDB"""
     config = configparser.ConfigParser()
     config.read(os.path.join(os.path.dirname(__file__),
-        '../docs/codes3d.conf'))
+                             '../docs/codes3d.conf'))
     db_dir = os.path.join(os.path.dirname(__file__), config.get('Defaults',
-        'CONSENSUS_DB_FP'))
+                                                                'CONSENSUS_DB_FP'))
 
     db_connect = sqlite3.connect(db_dir)
     db_cursor = db_connect.cursor()
 
     pathways = {}
-    query = (gene,)
+    query = (gene, 'KEGG', 'Reactome', 'Wikipathways')
     for entry in db_cursor.execute("""SELECT database, pathway_id, pathway_label
                                       FROM pathways
-                                      WHERE gene=?""", query):
+                                      WHERE gene=?
+                                      AND database NOT IN (?, ?, ?)""", query):
         entry = [subentry.encode('utf-8') for subentry in entry]
-        try:
-            pathways[entry[0]][entry[1]] = entry[2]
-        except KeyError:
-            pathways[entry[0]] = {}
-            pathways[entry[0]][entry[1]] = entry[2]
+        pathways[(entry[0], entry[1])] = entry[2]
 
     db_connect.close()
+
     return pathways
 
-def test(genes):
-    """Test functionality"""
-    pprinter = PrettyPrinter()
+def get_pathways(gene):
+    """All databases"""
+    pathways = {}
+    databases = (consensuspathdb, kegg, reactome, wikipathways)
 
-    build_consensus_db()
-    for gene in genes:
-        print("\nConsensusPathDB: %s" % gene)
-        pprinter.pprint(consensus_db(gene))
-        print("\nKEGG: %s" % gene)
-        pprinter.pprint(kegg(gene))
-        print("\nWikiPathways: %s" % gene)
-        pprinter.pprint(wikipathways(gene, False))
-        print("\nReactome: %s" % gene)
-        pprinter.pprint(reactome(gene))
-        print("\nReactome in WikiPathways: %s" % gene)
-        pprinter.pprint(reactome_in_wikipathways(gene))
+    for database in databases:
+        pathways.update(database(gene))
+
+    return pathways
 
 def main():
-    """Main"""
+    """Test functionality"""
     with open(argv[1]) as gene_file:
-        gene_identifiers = gene_file.read().splitlines()
-    test(gene_identifiers)
+        genes = gene_file.read().splitlines()
+
+    pprinter = PrettyPrinter()
+    build_consensuspathdb()
+
+    for gene in genes:
+        print(gene)
+        pprinter.pprint(get_pathways(gene))
+
+    remove_consensuspathdb()
 
 if __name__ == '__main__':
     main()
